requirements.txt
# Server Bot Requirements
# ======================

# Discord.py for Discord bot functionality
discord.py>=2.3.0

# Async support and utilities  
aiohttp>=3.8.0

# Logging and environment
python-dotenv>=0.19.0

# Date and time handling
python-dateutil>=2.8.0
# requests for API calls (if needed)
requests>=2.28.0

# Utility libraries
typing-extensions>=4.0.0